+++ src/core/server/api/consts/constDefaultSystems.ts
@@ -1,7 +1,9 @@
 import { DefaultDeathSystem } from '@AthenaServer/systems/defaults/death';
+import { DefaultTimeSystem } from '@AthenaServer/systems/defaults/time';
 import { DefaultWeatherSystem } from '@AthenaServer/systems/defaults/weather';
 
 export const defaultSystemsConst = {
     death: DefaultDeathSystem,
+    time: DefaultTimeSystem,
     weather: DefaultWeatherSystem,
 };